Create new app to log users in
A new Django app was created that will defined the views and
templates that enable users to log in to the admin site. Right
now, the website does have a login page, but it doesn't actually
allow users to log in. The functionality of the login_page view
function still needs to be written.

An end-to-end test was written that ensures that the superuser
can log in successfully. Again, because the functionality of the
login_page view hasn't been completed yet, this test fails.