Install SASS and Bootstrap
A package.json file was created to install Node packages and run
npm commands. Both SASS and Bootstrap were installed, and a
'npm build' command was defined in the package.json file that
compiles a .scss file to .css. The login.html template file was
modified to include this CSS file.