Refactor some files
The 'login' app was renamed to 'dashboard'. Furthermore, some
parts of the dashboard app were refactored: The unit tests were
moved to separate files within a 'tests' directory, and a new
decorator function was written that returns a 404 response if
a user is not authenticated. This new decorator will be used by
several view functions.