Configure Vite to compile TypeScript and SASS
Vite was added as a Node dev dependency. It has been configured to
compile all TypeScript files to JavaScript and SASS files to CSS.

Several changes were made to the way the Django templates are
rendered so that the correct React app files are used in
development versus in production.