Change dashboard modal React root ID
The id attribute of the div element that the dashboard modal
React app is mounted to was changed. This is important because,
in the future, the website may have multiple different React
apps mounted on the same page, and they can't all mount to an
element with an ID of "root".