Change get_vite_asset function
The get_vite_asset function that defines the functionality of the
vite_assets template tag was modified. It previously required the
STATIC_ROOT setting to be assigned a value. If it wasn't, an
exception was raised. This caused errors to be thrown when a lot
of the site's tests were run. Now, it uses the static files in
the 'static' directory if no STATIC_ROOT setting is defined. All
tests pass now (except for one that will be fixed later).