Enable modal to display proper links
The modal window that prompts users to enable 2FA was modified so
that it now also prompts users to verify their email address.
Furthermore, the proper links are now displayed in this modal
window. This ensures that users are redirected to the correct
page if they choose to enable two-factor authentication or verify
their email address.